goserbench: Make this package a module
This makes the goserbench package into its own module. This allows third
parties to import the main benchmark procedure without having to require
any of the tested serializers.

The goal of this split is to allow serializer projects to themselves
import the benchmark code to ease development and verification of the
benchmark numbers, without having to depend upon any of the other
serializers tested in this repository.

The main go.mod file is modified to ensure running the benchmarks
locally always uses the local version of the goserbench package. This
commit also tidies the go.mod and go.sum file.